pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "athenaforge"
version = "0.1.0"
description = "Searce AthenaForge™ — Athena-to-BigQuery Migration Accelerator"
requires-python = ">=3.11"
dependencies = [
    "pyyaml>=6.0",
    "click>=8.1",
    "rich>=13.0",
    "pydantic>=2.0",
    "mcp>=1.0",
    "networkx>=3.1",
    "jinja2>=3.1",
    "google-cloud-bigquery>=3.11",
    "google-cloud-bigquery-migration>=0.11",
    "google-cloud-storage>=2.10",
    "google-cloud-dlp>=3.12",
    "google-cloud-dataplex>=2.0",
    "boto3>=1.28",
    "sqlglot>=18.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-asyncio>=0.21",
    "pytest-cov>=4.1",
    "mypy>=1.5",
    "ruff>=0.1",
    "hypothesis>=6.82",
]

[project.scripts]
athenaforge = "athenaforge.presentation.cli.main:cli"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.coverage.run]
source = ["src/athenaforge"]

[tool.coverage.report]
show_missing = true
